requirements.txt
Django==4.0.5
uWSGI==2.0.20
pandas==2.2.3
openpyxl==3.1.5
psycopg2-binary==2.9.10